# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.CreateAdminRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      role:
        description: super_admin, support, moderator
        type: string
    type: object
...
  admin.UpdateAdminRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
      password:
        description: optional
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: API server for booking system.
  title: PhotoStudio API
  version: "1.0"
paths:
...
  /admin/management/admins:
    get:
      consumes:
      - application/json
      description: Get a paginated list of administrators
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List admins
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create a new administrator account
      parameters:
      - description: Admin fields
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/admin.CreateAdminRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create admin
      tags:
      - Admin
  /admin/management/admins/{id}:
    patch:
      consumes:
      - application/json
      description: Update an existing administrator account
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Admin fields
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update admin
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"